// node-sdk/src/openapi/models/AccountAllOfBranding.ts
/**
 * Pluto
 * The API for Pluto
 *
 * OpenAPI spec version: 1.0
 * Contact: support@pluto.co
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class AccountAllOfBranding {
    'logo'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "logo",
            "baseName": "logo",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AccountAllOfBranding.attributeTypeMap;
    }

    public constructor() {
    }
}