// node-sdk/src/openapi/models/AccountsBranding.ts
/**
 * Prism
 * The API for Prism
 *
 * OpenAPI spec version: 1.0
 * Contact: support@prism.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class AccountsBranding {
    'logo'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "logo",
            "baseName": "logo",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AccountsBranding.attributeTypeMap;
    }

    public constructor() {
    }
}